Updated ReadWrite UInt32 Unit Test Endian Check
A check had been added to the default BinaryStream test to ensure that
it used little endian by default. This check was invalid for two
reasons.

1. It assumes that the default endianness is little endian if none is
supplied. This is not true. It sets to the system's endianness. This has
been corrected to check that it is the system's endianness.

2. It only checked with the first value in the values array to keep the
check simple. This choice was somewhat arbitrary, and relies on the
first value not being symettrical, such as 0. If 0 is the value, the
test will pass regardless of endianness. Now the full array is checked.
Ideally this would be most effective with at least one non-symettrical
value.